+++ pyabc/epsilon/base.py
@@ -21,7 +21,9 @@ class Epsilon(ABC):
 
     def initialize(self,
                    t: int,
-                   get_weighted_distances: Callable[[], pd.DataFrame]):
+                   get_weighted_distances: Callable[[], pd.DataFrame],
+                   max_nr_populations: int,
+                   acceptor_config: dict):
         """
         This method is called by the ABCSMC framework before the first usage
         of the epsilon and can be used to calibrate it to the statistics of the
@@ -34,15 +36,20 @@ class Epsilon(ABC):
 
         t: int
             The time point to initialize the epsilon for.
-
         get_weighted_distances: Callable[[], pd.DataFrame]
             Returns on demand the distances for initializing the epsilon.
+        max_nr_populations: int
+            The maximum number of populations.
+        acceptor_config: dict
+            An object provided by the Acceptor class.
         """
         pass
 
     def update(self,
                t: int,
-               weighted_distances: pd.DataFrame):
+               weighted_distances: pd.DataFrame,
+               acceptance_rate: float,
+               acceptor_config: dict):
         """
         Update epsilon value to be used as acceptance criterion for
         generation t.
@@ -55,13 +62,16 @@ class Epsilon(ABC):
         t: int
             The generation index to update / set epsilon for. Counting is
             zero-based. So the first population has t=0.
-
         weighted_distances: pd.DataFrame
             The distances that should be used to update epsilon, as returned
             by Population.get_weighted_distances(). These are usually the
             distances of samples accepted in population t-1. The distances may
             differ from those used for acceptance in population t-1, if the
             distance function for population t has been updated.
+        acceptance_rate: float
+            The current generation's acceptance rate.
+        acceptor_config: dict
+            An object provided by the Acceptor class.
         """
         pass